GraphRAG/requirements.txt
# pip install --upgrade pip
python-dotenv
neo4j
pypdf
langchain
langchain-google-genai
langchain-community
langchain-experimental
# langchain-huggingface
transformers
sentence-transformers
json-repair
tiktoken
langchain_neo4j
openai